+++ src/main/java/command/admin/mapping/MappingCreateCommand.java
@@ -141,7 +141,7 @@ public class MappingCreateCommand implements Command {
         // 상위 카테고리
         String parentIdStr = request.getParameter("parentId");
         if (parentIdStr != null && !parentIdStr.isEmpty()) {
-            dto.setParentId(Integer.parseInt(parentIdStr));
+            dto.setParentId(Long.parseLong(parentIdStr));
         }
 
         // 순서
